Created databases.go and moved database specific interfaces there, out of core.go
